Removed type declaration
We had this break on the commandline, when called withoud website. This is probably because of another error where i removed the website from the customer. That broke somewhere.

But this does nog resolve the website to a default. And the website is not needed on ce because ce only uses 1 website.

Type Declarations breaks when a website is null.
 When called on a commandline the current website does nog give a website instance because it is not a frontend call, so i added the default website as fallback.

This is not a solution, just a patch probably.